+++ packages/terra-clinical-detail-view/src/terra-dev-site/test/clinical-detail-view/DetailView/DetailViewNoDivider.test.jsx
@@ -16,7 +16,7 @@ const DetailViewNoDivider = () => (
     secondaryTitles={['secondaryTitle1', 'secondaryTitle2']}
     subtitles={['Subtitle 1', 'Subtitle 2']}
     graph={(
-      <div className={cx('graph-content')}>
+      <div className={cx('graph-content')} tabIndex="0">
         This is where any visualizations would go
       </div>
 )}