+++ packages/terra-clinical-detail-view/CHANGELOG.md
@@ -6,6 +6,16 @@ Unreleased
 ### Changed
 * Axe violation fixed for scrollable content.
 
+3.10.0 - (September 26, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+3.9.0 - (September 19, 2019)
+----------
+### Changed
+* Convert SCSS files to .module.scss
+
 3.8.0 - (September 6, 2019)
 ----------
 ### Changed
@@ -13,6 +23,7 @@ Unreleased
 * Replaced Object.assign syntax with Object spread syntax
 * Add comments for ESlint 6
 * Cleaned up imports in examples and test files
+* Removed clinical-item-display as a dependency.
 
 3.7.0 - (August 14, 2019)
 ----------